chore(meta): move TxnGetResponse utils to separate file
